+++ src/ts/Errors/WebViewError.ts
@@ -1,0 +1,10 @@
+export class WebViewError {
+
+    public message?: string;
+    public name?: string;
+
+    constructor(message?: string, name?: string) {
+        this.message = message;
+        this.name = name;
+    }
+}