+++ src/ts/Test/Unit/AdUnits/VastAdUnitTest.ts
@@ -13,6 +13,7 @@ import { WakeUpManager } from 'Managers/WakeUpManager';
 import { Placement } from 'Models/Placement';
 import { AndroidVideoAdUnitController } from 'AdUnits/AndroidVideoAdUnitController';
 import { Platform } from 'Constants/Platform';
+import { VastEndScreen } from 'Views/VastEndScreen';
 
 import EventTestVast from 'xml/EventTestVast.xml';
 
@@ -63,10 +64,22 @@ describe('VastAdUnit', () => {
             const urlTemplate = 'http://foo.biz/%ZONE%/123';
             sandbox.stub(vast, 'getTrackingEventUrls').returns([ urlTemplate ]);
             sandbox.stub(eventManager, 'thirdPartyEvent').returns(null);
-            adUnit.sendTrackingEvent(eventManager, 'eventName', 'sessionId');
+            adUnit.sendTrackingEvent(eventManager, 'eventName', 'sessionId', 'sdkVersion');
 
             sinon.assert.calledOnce(<sinon.SinonSpy>eventManager.thirdPartyEvent);
             sinon.assert.calledWith(<sinon.SinonSpy>eventManager.thirdPartyEvent, 'vast eventName', 'sessionId', 'http://foo.biz/' + placement.getId() + '/123');
+        });
+
+        it('should replace "%SDK_VERSION%" in the url with the SDK version as a query parameter', () => {
+            const placement = adUnit.getPlacement();
+            const urlTemplate = 'http://ads-brand-postback.unityads.unity3d.com/brands/2002/defaultVideoAndPictureZone/%ZONE%/impression/common?adSourceId=2&advertiserDomain=appnexus.com&advertisingTrackingId=49f7acaa-81f2-4887-9f3b-cd124854879c&cc=USD&creativeId=54411305&dealCode=&demandSeatId=1&fillSource=appnexus&floor=0&gamerId=5834bc21b54e3b0100f44c92&gross=0&networkId=&precomputedFloor=0&seatId=958&value=1.01&sdkVersion=%SDK_VERSION%';
+            const vast = (<VastCampaign> adUnit.getCampaign()).getVast();
+            sandbox.stub(vast, 'getTrackingEventUrls').returns([ urlTemplate ]);
+            sandbox.stub(eventManager, 'thirdPartyEvent').returns(null);
+            adUnit.sendTrackingEvent(eventManager, 'start', 'sessionId', 'sdkVersion');
+
+            sinon.assert.calledOnce(<sinon.SinonSpy>eventManager.thirdPartyEvent);
+            sinon.assert.calledWith(<sinon.SinonSpy>eventManager.thirdPartyEvent, 'vast start', 'sessionId', 'http://ads-brand-postback.unityads.unity3d.com/brands/2002/defaultVideoAndPictureZone/' + placement.getId() + '/impression/common?adSourceId=2&advertiserDomain=appnexus.com&advertisingTrackingId=49f7acaa-81f2-4887-9f3b-cd124854879c&cc=USD&creativeId=54411305&dealCode=&demandSeatId=1&fillSource=appnexus&floor=0&gamerId=5834bc21b54e3b0100f44c92&gross=0&networkId=&precomputedFloor=0&seatId=958&value=1.01&sdkVersion=sdkVersion');
         });
     });
 
@@ -95,6 +108,15 @@ describe('VastAdUnit', () => {
 
             sinon.assert.calledOnce(<sinon.SinonSpy>eventManager.thirdPartyEvent);
             sinon.assert.calledWith(<sinon.SinonSpy>eventManager.thirdPartyEvent, 'vast impression', 'sessionId', 'http://foo.biz/sdkVersion/456');
+        });
+
+        it('should replace "%SDK_VERSION%" in the url with the SDK version as a query parameter', () => {
+            const urlTemplate = 'http://ads-brand-postback.unityads.unity3d.com/brands/2002/defaultVideoAndPictureZone/impression/common?adSourceId=2&advertiserDomain=appnexus.com&advertisingTrackingId=49f7acaa-81f2-4887-9f3b-cd124854879c&cc=USD&creativeId=54411305&dealCode=&demandSeatId=1&fillSource=appnexus&floor=0&gamerId=5834bc21b54e3b0100f44c92&gross=0&networkId=&precomputedFloor=0&seatId=958&value=1.01&sdkVersion=%SDK_VERSION%';
+            sandbox.stub(vast, 'getImpressionUrls').returns([ urlTemplate ]);
+            adUnit.sendImpressionEvent(eventManager, 'sessionId', 'sdkVersion');
+
+            sinon.assert.calledOnce(<sinon.SinonSpy>eventManager.thirdPartyEvent);
+            sinon.assert.calledWith(<sinon.SinonSpy>eventManager.thirdPartyEvent, 'vast impression', 'sessionId', 'http://ads-brand-postback.unityads.unity3d.com/brands/2002/defaultVideoAndPictureZone/impression/common?adSourceId=2&advertiserDomain=appnexus.com&advertisingTrackingId=49f7acaa-81f2-4887-9f3b-cd124854879c&cc=USD&creativeId=54411305&dealCode=&demandSeatId=1&fillSource=appnexus&floor=0&gamerId=5834bc21b54e3b0100f44c92&gross=0&networkId=&precomputedFloor=0&seatId=958&value=1.01&sdkVersion=sdkVersion');
         });
 
         it('should replace both "%ZONE%" and "%SDK_VERSION%" in the url with corresponding parameters', () => {
@@ -148,14 +170,14 @@ describe('VastAdUnit', () => {
         it('should call video click tracking url', () => {
             sandbox.stub(vast, 'getVideoClickTrackingURLs').returns(['https://www.example.com/foo/?bar=baz&inga=42&quux', 'http://wwww.tremor.com/click']);
             sandbox.stub(eventManager, 'thirdPartyEvent').returns(null);
-            adUnit.sendVideoClickTrackingEvent(eventManager, 'foo');
+            adUnit.sendVideoClickTrackingEvent(eventManager, 'foo', 'sdkVersion');
             sinon.assert.calledTwice(<sinon.SinonSpy>eventManager.thirdPartyEvent);
         });
 
         it('should not call thirdPartyEvent if there are no tracking urls', () => {
             sandbox.stub(vast, 'getVideoClickTrackingURLs').returns([]);
             sandbox.stub(eventManager, 'thirdPartyEvent').returns(null);
-            adUnit.sendVideoClickTrackingEvent(eventManager, 'foo');
+            adUnit.sendVideoClickTrackingEvent(eventManager, 'foo', 'sdkVersion');
             sinon.assert.notCalled(<sinon.SinonSpy>eventManager.thirdPartyEvent);
         });
     });
@@ -167,7 +189,7 @@ describe('VastAdUnit', () => {
             mockEventManager.expects('thirdPartyEvent').withArgs(`vast ${quartileEventName}`, '123', `http://localhost:3500/brands/14851/${quartileEventName}?advertisingTrackingId=123456&androidId=aae7974a89efbcfd&creativeId=CrEaTiVeId1&demandSource=tremor&gameId=14851&ip=192.168.69.69&token=9690f425-294c-51e1-7e92-c23eea942b47&ts=2016-04-21T20%3A46%3A36Z&value=13.1&zone=123`);
 
             const quartilePosition = campaign.getVast().getDuration() * 0.25 * quartile * 1000;
-            adUnit.sendProgressEvents(eventManager, '123', quartilePosition + 100, quartilePosition - 100);
+            adUnit.sendProgressEvents(eventManager, '123', 'sdkVersion', quartilePosition + 100, quartilePosition - 100);
             mockEventManager.verify();
         };
 
@@ -196,8 +218,52 @@ describe('VastAdUnit', () => {
             const mockEventManager = sinon.mock(eventManager);
             mockEventManager.expects('thirdPartyEvent').withArgs('vast video click', '123', 'http://myTrackingURL.com/click');
 
-            adUnit.sendVideoClickTrackingEvent(eventManager, '123');
+            adUnit.sendVideoClickTrackingEvent(eventManager, '123', 'sdkVersion');
             mockEventManager.verify();
+        });
+    });
+
+    describe('with companion ad', () => {
+        let vast: Vast;
+        let vastEndScreen: VastEndScreen;
+
+        beforeEach(() => {
+            vast = new Vast([], []);
+            const placement = TestFixtures.getPlacement();
+            campaign = new VastCampaign(vast, 'campaignId', 'gamerId', 12);
+            const overlay = <Overlay><any> sinon.createStubInstance(Overlay);
+            const nativeBridge = TestFixtures.getNativeBridge();
+            const androidVideoAdUnitController = new AndroidVideoAdUnitController(nativeBridge, TestFixtures.getDeviceInfo(Platform.ANDROID), placement, campaign, overlay, null);
+            vastEndScreen = <VastEndScreen><any> {
+                hide: sinon.spy(),
+                remove: sinon.spy()
+            };
+            adUnit = new VastAdUnit(nativeBridge, androidVideoAdUnitController, vastEndScreen);
+        });
+
+        it('should return correct companion click through url', () => {
+            sandbox.stub(vast, 'getCompanionClickThroughUrl').returns('http://www.example.com/wpstyle/?p=364');
+
+            const clickThroughURL = adUnit.getCompanionClickThroughUrl();
+            assert.equal(clickThroughURL, 'http://www.example.com/wpstyle/?p=364');
+        });
+
+        it('should return null when companion click through url is invalid', () => {
+            sandbox.stub(vast, 'getCompanionClickThroughUrl').returns('blah');
+
+            const clickThroughURL = adUnit.getCompanionClickThroughUrl();
+            assert.equal(clickThroughURL, null);
+        });
+
+        it('should return endscreen', () => {
+            const endScreen = adUnit.getEndScreen();
+            assert.equal(endScreen, vastEndScreen);
+        });
+
+        it('should hide and then remove endscreen on hide', () => {
+            adUnit.hide();
+            sinon.assert.called(<sinon.SinonSpy>vastEndScreen.hide);
+            sinon.assert.called(<sinon.SinonSpy>vastEndScreen.remove);
         });
     });
 });