+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# Mon Dec 19 13:33:52 UTC 2016
+
+* Add campaign ID to video_too_long diagnostic message
+* Add test option to force a country, campaign or A/B group
+* Send only Google Play Advertising ID or Android ID in diagnostic messages, never both
+* Add Xmas endscreen to all A/B groups except 8 and 9
+
+# Thu Dec 15 16:46:43 UTC 2016
+
+* Removed Marvel AB test from groups 8-11
+* Added christmas theme AB test to groups 8 & 9
+
+# Tue Dec 13 11:49:18 UTC 2016
+
+* Diagnostics message for videos over 40 seconds
+
+# Mon Dec 12 15:55:37 UTC 2016
+
+* Update to TypeScript 2.1.4
+* Update dependencies to latest
+* Add SDK version to all VAST tracking/impression urls
+
+# Thu Dec  8 15:37:41 UTC 2016
+
+* Optional end screen support for brand ads
+* Rating count abbreviation support
+
+# Wed Dec  7 16:48:28 UTC 2016
+
+* Disable fake landscape introduced in 2.0.6 due to iOS native bug
+* Marvel Contest of Champions custom animated endscreens (A/B groups from 8 to 11)
+* Restore accidentally removed rules for endscreen portrait and landscape image assets
+
 # Wed Nov 30 13:40:07 UTC 2016
 
 * Deploy playable ads to all users except for A/B groups 6 and 7
@@ -17,3 +50,4 @@
 * If caching fails due to network error, retry five times with a 10 second delay between retries
 * Fix iOS cache retries, previously they never retried if downloading failed due to network error
 * Disable hardware acceleration on one Android 4.2 device (ABT-91)
+